backend/requirements.txt
# Web-API
Flask~=3.0.3
flask-cors~=5.0.0

# Datenbank (enthält gridfs)
pymongo~=4.13.0

# Sicherheit und Login
bcrypt==3.2.0
PyJWT~=2.9.0

# KI und Chatbot
torch~=2.7.0
transformers~=4.40.2
accelerate
thefuzz[speedup]~=0.22.1

# Bildverarbeitung
ultralytics~=8.3.115
opencv-python
scikit-learn
joblib~=1.4.2
Pillow~=11.0.0
certifi~=2024.2.2
jwt~=1.3.1
numpy~=2.1.1
requests~=2.31.0